requirements.txt
asyncio
anyio
pydantic
pytest
redis
nats-py
rich
python-ulid
aiohttp==3.11.10
thorpy
numpy
pillow
coolname
pandas
pytest-cov
pyparsing
pyyaml